Update README.rst
When I was attempting to follow these instructions, I got an error that version 1.8 was not installed and Otto pointed out that it was version 11 that we just installed.  Changing the command to have -v 11 worked.  I am not sure where the 1.8 came from and wondered if possibly it was meant to be 11.8.  This fix has 11.